smartbot/requirements.txt
google-adk
google-generativeai==0.8.3
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic==2.9.2
websockets==13.0.1
